Fixing issue 725 error messages fails to start correctly when metadata is not available
