+++ internal/support/scheduler/init.go
@@ -189,7 +189,6 @@ func setLoggingTarget() string {
 
 func initializeClients(useConsul bool) {
 	// Create metadata clients
-
 	params := types.EndpointParams{
 		ServiceKey:  internal.SupportSchedulerServiceKey,
 		Path:        clients.ApiScheduleRoute,
@@ -208,5 +207,5 @@ func initializeClients(useConsul bool) {
 	// metadata Addressable client
 	params.Path = clients.ApiAddressableRoute
 	params.Url = Configuration.Clients["Metadata"].Url() + clients.ApiAddressableRoute
-	mac = metadata.NewAddressableClient(params,startup.Endpoint{})
+	mac = metadata.NewAddressableClient(params, startup.Endpoint{})
 }
